requirements.txt
accelerate==1.0.1
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
albucore==0.0.17
albumentations==1.4.18
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.5.2
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.4
async-timeout==5.0.1
attrs==25.3.0
autoattack @ file:///gemini/code/temp/auto-attack-master
babel==2.17.0
backcall==0.2.0
backports.tarfile==1.2.0
beautifulsoup4==4.13.4
bleach==6.1.0
blessed==1.21.0
braceexpand==0.1.7
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
chardet==4.0.0
charset-normalizer==3.4.2
click==8.1.8
clip-interrogator==0.6.0
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.1.1
cryptography==45.0.3
cssselect==1.2.0
cssutils==2.11.1
cycler==0.12.1
Cython==3.2.0
datasets==3.1.0
datclass==0.2.28
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.3.1
dill==0.3.8
docopt==0.6.2
docutils==0.20.1
einops==0.8.1
eval_type_backport==0.2.2
exceptiongroup==1.3.0
executing==2.2.0
fastapi==0.121.1
fastjsonschema==2.21.1
ffmpy==0.5.0
filelock==3.16.1
fonttools==4.57.0
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.9.0
ftfy==6.2.3
fvcore==0.1.5.post20221221
gdown==5.2.0
geotorch @ file:///gemini/code/temp/geotorch-master
gpustat==1.1.1
GPUtil==1.4.0
grad-cam @ file:///gemini/code/temp/pytorch-grad-cam-master
gradio==4.44.1
gradio_client==1.3.0
h11==0.16.0
h5py==3.11.0
hf-xet==1.1.3
hopcroftkarp==1.2.5
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.32.4
humanfriendly==10.0
id==1.5.0
idna==2.10
imageio==2.35.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
iopath==0.1.10
ipykernel==6.29.5
ipython==8.12.3
ipywidgets==8.1.7
isoduration==20.11.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jedi==0.19.2
jeepney==0.9.0
Jinja2==3.1.6
joblib==1.4.2
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.3.7
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
keyring==25.5.0
kiwisolver==1.4.7
lazy_loader==0.4
littleutils==0.2.4
loguru==0.7.3
lor-deckcodes==5.0.0
lxml==6.0.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
more-itertools==10.5.0
multidict==6.1.0
multiprocess==0.70.16
nbclient==0.10.1
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.1
nh3==0.2.21
notebook==7.3.3
notebook_shim==0.2.4
numpy==1.24.4
nvidia-ml-py==12.575.51
nvitop==1.5.1
ogb==1.3.6
open_clip_torch==2.32.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
orjson==3.10.15
outcome==1.3.0.post0
outdated==0.2.2
overrides==7.7.0
packaging==25.0
pandas==1.2.3
pandocfilters==1.5.1
parso==0.8.4
persim==0.3.8
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pkgutil_resolve_name==1.3.10
platformdirs==4.3.6
portalocker==3.0.0
POT==0.9.5
pqi==4.0.2
prefetch_generator==1.0.3
premailer==3.10.0
prettytable==3.11.0
prometheus_client==0.21.1
prompt_toolkit==3.0.51
propcache==0.2.0
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.1
pyloco==0.0.139
pyot==6.0.9
pyparsing==3.1.4
pypng==0.20220715.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-json-logger==3.3.0
python-multipart==0.0.20
pytz==2025.2
PyWavelets==1.4.1
PyYAML==6.0.2
pyzmq==26.4.0
qrcode==7.4.2
qrcode-terminal==0.8
readme_renderer==43.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986==2.0.0
rfc3986-validator==0.1.1
rich==14.0.0
ripser==0.6.12
rpds-py==0.20.1
ruff==0.14.4
safetensors==0.5.3
schedule==1.2.2
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
SecretStorage==3.3.3
selenium==4.27.1
semantic-version==2.10.0
Send2Trash==1.8.3
setproctitle==1.3.6
shellingham==1.5.4
SimpleWebSocketServer==0.1.2
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
stack-data==0.6.3
starlette==0.44.0
tabulate==0.9.0
termcolor==2.4.0
terminado==0.18.1
threadpoolctl==3.5.0
tifffile==2023.7.10
tinycss2==1.2.1
tokenizers==0.15.2
tomli==2.2.1
tomlkit==0.12.0
torch==1.13.1+cu117
torch_scatter @ file:///gemini/code/temp/pytorch_scatter-2.1.0
torchaudio==0.13.1+cu117
torchdiffeq==0.2.5
torchstat2==0.0.0
torchvision==0.14.1+cu117
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.37.2
trio==0.27.0
trio-websocket==0.12.2
ttach==0.0.3
twine==6.1.0
typer==0.20.0
types-python-dateutil==2.9.0.20241206
typing==3.7.4.3
typing_extensions==4.13.2
uri-template==1.3.0
urllib3==2.2.3
ushlex==0.99.1
uvicorn==0.33.0
wcwidth==0.2.13
webcolors==24.8.0
webdataset==0.2.100
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
wget==3.2
widgetsnbextension==4.0.14
wilds==2.0.0
wrapt==2.0.1
wsproto==1.2.0
xxhash==3.6.0
yacs==0.1.8
yagmail==0.15.293
yarl==1.15.2
zipp==3.20.2
